Restructure JSON serialization logic
